% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{equal_breaks}
\alias{equal_breaks}
\title{Equal breaks for ggplot2 facets}
\usage{
equal_breaks(n = 3, s = 0.05)
}
\arguments{
\item{n}{Number of breaks}

\item{s}{Scaling factor (cf. multiplicative expand)}
}
\description{
Equal breaks for ggplot2 facets
}
\details{
Code copied from: https://stackoverflow.com/questions/28436855/change-the-number-of-breaks-using-facet-grid-in-ggplot2
}
\examples{

library(ggplot2)

ggplot(mtcars, aes(x = disp, y = drat)) +
  geom_point() +
  facet_wrap(~cyl, scales = "free_x") +
  scale_x_continuous(breaks = equal_breaks(n = 4))
}
