% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpmthemes-package.R
\docType{package}
\name{ggpmthemes-package}
\alias{ggpmthemes}
\alias{ggpmthemes-package}
\title{ggpmthemes: Personal Themes for ggplot2}
\description{
Compilation of different ggplot2 themes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PMassicotte/ggpmthemes}
  \item Report bugs at \url{https://github.com/PMassicotte/ggpmthemes/issues}
}

}
\author{
\strong{Maintainer}: Philippe Massicotte \email{pmassicotte@hotmail.com}

}
\keyword{internal}
