% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scientific_10x}
\alias{scientific_10x}
\title{Display scientific notation labels}
\usage{
scientific_10x(x)
}
\arguments{
\item{x}{A number to be displayed as scientific notation.}
}
\description{
Display scientific notation labels
}
\examples{
library(ggplot2)

ggplot(mtcars, aes(x = disp, y = drat)) +
  geom_point() +
  facet_wrap(~cyl, scales = "free_x") +
  scale_x_continuous(labels = scientific_10x)
}
