% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exo.R
\name{theme_exo}
\alias{theme_exo}
\title{A ggplot2 theme based on the Google Exo font.}
\usage{
theme_exo(base_size = 12, base_family = "Exo",
  base_line_size = base_size/22, base_rect_size = base_size/22)
}
\arguments{
\item{base_size}{Font size. Default at 12 pts.}

\item{base_family}{Font family. Default is "Exo"}

\item{base_line_size}{Base line size}

\item{base_rect_size}{Base rex size}
}
\description{
A ggplot2 theme based on the Google Exo font.
}
