% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_light_modified.R
\name{theme_light_modified}
\alias{theme_light_modified}
\title{A ggplot2 theme light modified.}
\usage{
theme_light_modified(
  base_size = 12,
  base_family,
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{Font size. Default at 12 pts.}

\item{base_family}{Font family. Default is "Exo"}

\item{base_line_size}{Base line size}

\item{base_rect_size}{Base rect size}
}
\description{
A ggplot2 theme light modified.
}
