% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maven_pro.R
\name{theme_maven}
\alias{theme_maven}
\title{A ggplot2 theme based on the Google Maven Pro font.}
\usage{
theme_maven(base_size = 12, base_family = "Maven Pro",
  base_line_size = base_size/22, base_rect_size = base_size/22)
}
\arguments{
\item{base_size}{Font size. Default at 12 pts.}

\item{base_family}{Font family. Default is "Maven Pro"}

\item{base_line_size}{Base line size}

\item{base_rect_size}{Base rect size}
}
\description{
A ggplot2 theme based on the Google Maven Pro font.
}
