% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montserrat.R
\name{theme_montserrat}
\alias{theme_montserrat}
\title{A ggplot2 theme based on the Google Montserrat font.}
\usage{
theme_montserrat(
  base_size = 12,
  base_family = "Montserrat",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{Font size. Default at 12 pts.}

\item{base_family}{Font family. Default is "Montserrat"}

\item{base_line_size}{Base line size}

\item{base_rect_size}{Base rect size}
}
\description{
A ggplot2 theme based on the Google Montserrat font.
}
