% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poppins.R
\name{theme_poppins}
\alias{theme_poppins}
\title{A ggplot2 theme based on the Google Poppins font.}
\usage{
theme_poppins(base_size = 12, base_family = "Poppins",
  base_line_size = base_size/22, base_rect_size = base_size/22)
}
\arguments{
\item{base_size}{Font size. Default at 12 pts.}

\item{base_family}{Font family. Default is "Poppins"}

\item{base_line_size}{Base line size}

\item{base_rect_size}{Base rect size}
}
\description{
A ggplot2 theme based on the Google Poppins font.
}
