% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpmthemes-package.R
\docType{package}
\name{ggpmthemes-package}
\alias{ggpmthemes}
\alias{ggpmthemes-package}
\title{ggpmthemes: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (YOUR-ORCID-ID)

}
\keyword{internal}
